Single type for client
