+++ crates/aranya-daemon/src/daemon.rs
@@ -26,6 +26,7 @@ use crate::{
     actions::Actions,
     api::{ApiKey, DaemonApiServer, PublicApiKey},
     aqc::Aqc,
+    aranya,
     config::Config,
     keystore::{AranyaStore, LocalStore},
     policy,
@@ -45,9 +46,8 @@ pub(crate) type SP = LinearStorageProvider<FileManager>;
 /// EF = Policy Effect
 pub(crate) type EF = policy::Effect;
 
-pub(crate) type TcpSyncClient = crate::sync::tcp::Client<EN, SP, CE>;
+pub(crate) type Client = aranya::Client<EN, SP, CE>;
 type TcpSyncServer = crate::sync::tcp::Server<EN, SP>;
-pub(crate) type ActionsClient = crate::actions::Client<EN, SP, CE>;
 
 /// The daemon itself.
 pub struct Daemon {
@@ -106,9 +106,6 @@ impl Daemon {
             (client, local_addr)
         };
 
-        // Initialize Aranya actions client.
-        let actions_client = ActionsClient::new(Arc::clone(&client.aranya));
-
         // Sync in the background at some specified interval.
         let (send_effects, recv_effects) = tokio::sync::mpsc::channel(256);
         let (mut syncer, peers) = Syncer::new(Arc::clone(&client), send_effects);
@@ -134,7 +131,7 @@ impl Daemon {
                 let mut peers = BTreeMap::new();
                 for graph_id in &graph_ids {
                     let graph_peers = BiBTreeMap::from_iter(
-                        actions_client
+                        client
                             .actions(graph_id)
                             .query_aqc_network_names_off_graph()
                             .await?,
@@ -147,7 +144,7 @@ impl Daemon {
         };
 
         let api = DaemonApiServer::new(
-            actions_client.into(),
+            client,
             local_addr,
             self.cfg.uds_api_path.clone(),
             api_sk,
@@ -185,7 +182,7 @@ impl Daemon {
         store: AranyaStore<KS>,
         pk: &PublicKeys<CS>,
         external_sync_addr: Addr,
-    ) -> Result<(TcpSyncClient, TcpSyncServer)> {
+    ) -> Result<(Client, TcpSyncServer)> {
         let device_id = pk.ident_pk.id()?;
 
         let aranya = Arc::new(Mutex::new(ClientState::new(
@@ -196,7 +193,7 @@ impl Daemon {
             ),
         )));
 
-        let client = TcpSyncClient::new(Arc::clone(&aranya));
+        let client = Client::new(Arc::clone(&aranya));
 
         let server = {
             info!(addr = %external_sync_addr, "starting TCP server");
