+++ crates/aranya-daemon/test-util/src/lib.rs
@@ -13,7 +13,8 @@ use aranya_crypto::{
     Csprng, Rng,
 };
 use aranya_daemon::{
-    actions::{self, Actions},
+    actions::Actions,
+    aranya,
     policy::{Effect, KeyBundle as DeviceKeyBundle, Role},
     sync,
     vm_policy::{PolicyEngine, VecSink, TEST_POLICY_1},
@@ -33,7 +34,7 @@ use tokio::{
 };
 
 // Aranya sync client for testing.
-pub type TestClient = sync::tcp::Client<
+pub type TestClient = aranya::Client<
     PolicyEngine<DefaultEngine, Store>,
     LinearStorageProvider<FileManager>,
     DefaultEngine,
@@ -42,13 +43,6 @@ pub type TestClient = sync::tcp::Client<
 // Aranya sync server for testing.
 pub type TestServer =
     sync::tcp::Server<PolicyEngine<DefaultEngine, Store>, LinearStorageProvider<FileManager>>;
-
-// Aranya actions client for testing.
-pub type TestActionsClient = actions::Client<
-    PolicyEngine<DefaultEngine, Store>,
-    LinearStorageProvider<FileManager>,
-    DefaultEngine,
->;
 
 // checks if effects vector contains a particular type of effect.
 #[macro_export]
@@ -61,8 +55,6 @@ macro_rules! contains_effect {
 pub struct TestDevice {
     /// Aranya sync client.
     aranya: TestClient,
-    /// Aranya actions client.
-    actions: TestActionsClient,
     /// The Aranya graph ID.
     pub graph_id: GraphId,
     /// The address that the server is listening on.
@@ -77,7 +69,6 @@ impl TestDevice {
     pub fn new(
         client: TestClient,
         server: TestServer,
-        actions: TestActionsClient,
         local_addr: Addr,
         pk: PublicKeys<DefaultCipherSuite>,
         graph_id: GraphId,
@@ -85,7 +76,6 @@ impl TestDevice {
         let handle = task::spawn(async { server.serve().await }).abort_handle();
         Ok(Self {
             aranya: client,
-            actions,
             graph_id,
             local_addr,
             handle,
@@ -97,7 +87,7 @@ impl TestDevice {
 impl TestDevice {
     pub async fn sync(&self, device: &TestDevice) -> Result<Vec<Effect>> {
         let mut sink = VecSink::new();
-        self.sync_peer(self.graph_id, &mut sink, &device.local_addr)
+        self.sync_peer_tcp(self.graph_id, &mut sink, &device.local_addr)
             .await
             .with_context(|| format!("unable to sync with peer at {}", device.local_addr))?;
         Ok(sink.collect()?)
@@ -110,7 +100,7 @@ impl TestDevice {
         LinearStorageProvider<FileManager>,
         DefaultEngine<Rng>,
     > {
-        self.actions.actions(&self.graph_id)
+        self.aranya.actions(&self.graph_id)
     }
 }
 
@@ -179,7 +169,7 @@ impl TestCtx {
         let root = self.dir.path().join(name);
         assert!(!root.try_exists()?, "duplicate client name: {name}");
 
-        let (client, server, actions, local_addr, pk) = {
+        let (client, server, local_addr, pk) = {
             let mut store = {
                 let path = root.join("keystore");
                 fs::create_dir_all(&path)?;
@@ -213,12 +203,14 @@ impl TestCtx {
 
             let aranya = Arc::new(Mutex::new(graph));
             let client = TestClient::new(Arc::clone(&aranya));
-            let server = TestServer::new(Arc::clone(&aranya), listener);
-            let actions = TestActionsClient::new(Arc::clone(&aranya));
-            (client, server, actions, local_addr, pk)
+            let server: sync::tcp::Server<
+                PolicyEngine<DefaultEngine, Store>,
+                LinearStorageProvider<FileManager>,
+            > = TestServer::new(Arc::clone(&aranya), listener);
+            (client, server, local_addr, pk)
         };
 
-        TestDevice::new(client, server, actions, local_addr.into(), pk, id)
+        TestDevice::new(client, server, local_addr.into(), pk, id)
     }
 
     /// Creates `n` members.
@@ -242,7 +234,7 @@ impl TestCtx {
                 let nonce = &mut [0u8; 16];
                 Rng.fill_bytes(nonce);
                 (client.graph_id, _) = client
-                    .actions
+                    .aranya
                     .create_team(DeviceKeyBundle::try_from(&client.pk)?, Some(nonce))
                     .await?;
             }
